# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Cloudfront",
    "description": "Cloudfront api endpoints",
    "version": "1.0.0"
  },
  "host": "localhost:8001",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/faqs/createFaq": {
      "post": {
        "tags": [
          "faqs"
        ],
        "description": "This endpoint creates a new FAQ",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "question": {
                  "example": "any"
                },
                "answer": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "FAQ created successfully"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/faqs/getFaqs": {
      "get": {
        "tags": [
          "faqs"
        ],
        "description": "This endpoint gets all FAQs",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "FAQs retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/faqs/updateFaq/{id}": {
      "put": {
        "tags": [
          "faqs"
        ],
        "description": "This endpoint updates a FAQ",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "question": {
                  "example": "any"
                },
                "answer": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "FAQ updated successfully"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "FAQ not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/faqs/deleteFaq/{id}": {
      "delete": {
        "tags": [
          "faqs"
        ],
        "description": "This endpoint deletes a FAQ",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "FAQ deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "FAQ not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  }
}